refactor(admin/components): move columns css from shared to requesttable
